backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pinecone-client
python-multipart
pydantic
